+++ backend/src/main/java/org/dsr/practice/services/BillsService.java
@@ -2,13 +2,13 @@ package org.dsr.practice.services;
 
 import org.dsr.practice.models.Account;
 import org.dsr.practice.models.Bill;
-import org.dsr.practice.models.BillInfo;
 import org.dsr.practice.models.User;
 import org.dsr.practice.repos.BillRepository;
 import org.dsr.practice.utils.PriceConverter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.stream.Collectors;
 
 @Service
@@ -50,7 +50,7 @@ public class BillsService {
         return billInfoService;
     }
 
-    public boolean Repayment(String pr, Long accountId, Long billId, Double value){
+    public boolean Repayment(String pr, Long accountId, Long billId, BigDecimal value){
         try {
             var billInfo = usersService
                     .getUser(pr)
